+++ car.h
@@ -30,7 +30,7 @@ int sleepm(long msec);
 int drive(int left, int right);
 
 // set speed for left and right motor, from -90 up to 90. timer specifies milliseconds untill stop
-int driveWithTimer(int left, int right, int timer);
+int driveWithTimer(int left, int right, long timer);
 
 // turns servor for the radar to look at the specified angle
 int look(int angle);
